INS-2216: move future and handler to separate package
