+++ network/transport/future/manager.go
@@ -48,71 +48,49 @@
 //    whether it competes with the products or services of Insolar Technologies GmbH.
 //
 
-package transport
+package future
 
 import (
-	"context"
+	"sync"
 
-	"github.com/insolar/insolar/instrumentation/inslogger"
-	"github.com/insolar/insolar/metrics"
+	"github.com/insolar/insolar/network"
 	"github.com/insolar/insolar/network/hostnetwork/packet"
-	"github.com/insolar/insolar/network/hostnetwork/packet/types"
 )
 
-type packetHandlerImpl struct {
-	futureManager futureManager
-
-	received chan *packet.Packet
+type futureManagerImpl struct {
+	mutex   sync.RWMutex
+	futures map[network.RequestID]Future
 }
 
-func newPacketHandlerImpl(futureManager futureManager) *packetHandlerImpl {
-	return &packetHandlerImpl{
-		futureManager: futureManager,
-		received:      make(chan *packet.Packet),
+func newFutureManagerImpl() *futureManagerImpl {
+	return &futureManagerImpl{
+		futures: make(map[network.RequestID]Future),
 	}
 }
 
-func (ph *packetHandlerImpl) Handle(ctx context.Context, msg *packet.Packet) {
-	metrics.NetworkPacketReceivedTotal.WithLabelValues(msg.Type.String()).Inc()
-	if msg.IsResponse {
-		ph.processResponse(ctx, msg)
-		return
-	}
+func (fm *futureManagerImpl) Create(msg *packet.Packet) Future {
+	future := NewFuture(msg.RequestID, msg.Receiver, msg, func(f Future) {
+		fm.delete(f.ID())
+	})
 
-	ph.processRequest(ctx, msg)
+	fm.mutex.Lock()
+	defer fm.mutex.Unlock()
+
+	fm.futures[msg.RequestID] = future
+
+	return future
 }
 
-func (ph *packetHandlerImpl) Received() <-chan *packet.Packet {
-	return ph.received
+func (fm *futureManagerImpl) Get(msg *packet.Packet) Future {
+	fm.mutex.RLock()
+	defer fm.mutex.RUnlock()
+
+	return fm.futures[msg.RequestID]
 }
 
-func (ph *packetHandlerImpl) processResponse(ctx context.Context, msg *packet.Packet) {
-	logger := inslogger.FromContext(ctx)
+func (fm *futureManagerImpl) delete(id network.RequestID) {
+	fm.mutex.Lock()
+	defer fm.mutex.Unlock()
 
-	logger.Debugf("[ processResponse ] Process response %s from %s with RequestID = %d", msg.Type, msg.RemoteAddress, msg.RequestID)
-
-	future := ph.futureManager.Get(msg)
-	if future != nil {
-		if shouldProcessPacket(future, msg) {
-			logger.Debugf("[ processResponse ] Processing future with RequestID = %v", msg.RequestID)
-			future.SetResult(msg)
-		} else {
-			logger.Debugf("[ processResponse ] Canceling future with RequestID = %v", msg.RequestID)
-		}
-		future.Cancel()
-	}
-}
-
-func (ph *packetHandlerImpl) processRequest(ctx context.Context, msg *packet.Packet) {
-	logger := inslogger.FromContext(ctx)
-	logger.Debugf("[ processRequest ] Process request %s from %s with RequestID = %d", msg.Type, msg.RemoteAddress, msg.RequestID)
-
-	ph.received <- msg
-}
-
-func shouldProcessPacket(future Future, msg *packet.Packet) bool {
-	typesShouldBeEqual := msg.Type == future.Request().Type
-	responseIsForRightSender := future.Actor().Equal(*msg.Sender)
-
-	return typesShouldBeEqual && (responseIsForRightSender || msg.Type == types.Ping)
+	delete(fm.futures, id)
 }
